Add files via upload
To design a temperature measuring system using a 1N4001 diode. The temperature 
is calculated and saved in the EEPROM to keep it safe. The temperature values 
is presented on an LCD together with the 2-minute timer. Additionally, a 
pushbutton must be implemented to read the last 15 temperature values and display this 
on the LCD screen.